requirements.txt
Kivy==2.2.1
kivymd==1.2.0
libusb_package==1.0.26.2
matplotlib==3.7.2
minimalmodbus==2.1.1
mysql_connector_repackaged==0.3.1
numpy==1.25.1
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
PyAudio==0.2.14
pymodbus==3.5.4
pyserial==3.5
python_escpos==3.1
python_snap7==2.0.0
pyusb==1.2.1
pywin32==306
scipy==1.14.1
